api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
passlib
bcrypt==4.0.1
python-jose[cryptography]
cryptography
email-validator
python-multipart